Restore time-based page revision IDs, but with hi-res time to avoid conflicts

+++ nlw/etc/socialtext/db/socialtext-schema.sql
@@ -1108,7 +1108,7 @@ CREATE TABLE page_link (
 CREATE TABLE page_revision (
     workspace_id bigint NOT NULL,
     page_id text NOT NULL,
-    revision_id bigint NOT NULL,
+    revision_id NUMERIC(20,6) NOT NULL,
     revision_num integer NOT NULL,
     name text,
     editor_id bigint NOT NULL,
